新增题库相关类，重新配置日志框架

+++ me-site/src/main/java/com/me/mesite/infrastructure/gatawayimpl/database/repository/UpmsSysUserRepository.java
@@ -3,6 +3,6 @@ package com.me.mesite.infrastructure.gatawayimpl.database.repository;
 import com.me.mesite.infrastructure.gatawayimpl.database.dataobject.UpmsSUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UpmsSysUserRepository extends JpaRepository<UpmsSUser, Long> {
+public interface UpmsSysUserRepository extends JpaRepository<UpmsSUser, Integer> {
 
 }
